Rettet småfeil og la til prototype

+++ class_note.cpp
@@ -3,7 +3,7 @@
 
 using namespace music;
 
-note::note(std::string c_key, double c_length){	
+note::note(std::string c_key, int c_length){	
 key = c_key;
 length = c_length;
 }
@@ -12,6 +12,6 @@ std::string note::getKey(){
 return key;
 }
 
-double note::getLength(){
+int note::getLength(){
 return length;
 }
